refactor: Reorganize project directory structure
