+++ src/main/java/com/sogonsogon/neighclova/dto/request/auth/EmailCheckRequestDto.java
@@ -1,4 +1,4 @@
-package com.sogonsogon.neighclova.dto.request;
+package com.sogonsogon.neighclova.dto.request.auth;
 
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
@@ -8,10 +8,8 @@ import lombok.Setter;
 @Getter
 @Setter
 @NoArgsConstructor
-public class SignInRequestDto {
-    @NotBlank
-    private String email;
+public class EmailCheckRequestDto {
 
     @NotBlank
-    private String password;
+    private String email;
 }
