+++ src/main/java/com/sogonsogon/neighclova/dto/response/auth/EmailCheckResponseDto.java
@@ -1,7 +1,8 @@
-package com.sogonsogon.neighclova.dto.response;
+package com.sogonsogon.neighclova.dto.response.auth;
 
 import com.sogonsogon.neighclova.common.ResponseCode;
 import com.sogonsogon.neighclova.common.ResponseMessage;
+import com.sogonsogon.neighclova.dto.response.ResponseDto;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
