+++ src/main/java/com/sogonsogon/neighclova/dto/response/auth/SignUpResponseDto.java
@@ -1,11 +1,10 @@
-package com.sogonsogon.neighclova.dto.response;
+package com.sogonsogon.neighclova.dto.response.auth;
 
 import com.sogonsogon.neighclova.common.ResponseCode;
 import com.sogonsogon.neighclova.common.ResponseMessage;
-import com.sogonsogon.neighclova.dto.request.SignUpRequestDto;
+import com.sogonsogon.neighclova.dto.response.ResponseDto;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
-import org.springframework.http.HttpStatusCode;
 import org.springframework.http.ResponseEntity;
 
 @Getter
