+++ src/main/java/com/sogonsogon/neighclova/dto/response/place/GetAllPlaceResponseDto.java
@@ -1,8 +1,9 @@
-package com.sogonsogon.neighclova.dto.response;
+package com.sogonsogon.neighclova.dto.response.place;
 
 import com.sogonsogon.neighclova.common.ResponseCode;
 import com.sogonsogon.neighclova.common.ResponseMessage;
 import com.sogonsogon.neighclova.dto.object.PlaceListItem;
+import com.sogonsogon.neighclova.dto.response.ResponseDto;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -10,7 +11,7 @@ import org.springframework.http.ResponseEntity;
 import java.util.List;
 
 @Getter
-public class GetAllPlaceResponseDto extends ResponseDto{
+public class GetAllPlaceResponseDto extends ResponseDto {
     private List<PlaceListItem> placeList;
 
     public GetAllPlaceResponseDto(List<PlaceListItem> PlaceListItem) {
