+++ src/main/java/com/sogonsogon/neighclova/dto/response/place/PlaceResponseDto.java
@@ -1,13 +1,14 @@
-package com.sogonsogon.neighclova.dto.response;
+package com.sogonsogon.neighclova.dto.response.place;
 
 import com.sogonsogon.neighclova.common.ResponseCode;
 import com.sogonsogon.neighclova.common.ResponseMessage;
+import com.sogonsogon.neighclova.dto.response.ResponseDto;
 import lombok.Getter;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 
 @Getter
-public class PlaceResponseDto extends ResponseDto{
+public class PlaceResponseDto extends ResponseDto {
 
     private PlaceResponseDto() {
         super(ResponseCode.SUCCESS, ResponseMessage.SUCCESS);
