+++ src/main/java/com/sogonsogon/neighclova/service/AuthService.java
@@ -1,7 +1,7 @@
 package com.sogonsogon.neighclova.service;
 
-import com.sogonsogon.neighclova.dto.request.*;
-import com.sogonsogon.neighclova.dto.response.*;
+import com.sogonsogon.neighclova.dto.request.auth.*;
+import com.sogonsogon.neighclova.dto.response.auth.*;
 import org.springframework.http.ResponseEntity;
 
 public interface AuthService {
