+++ src/main/java/com/sogonsogon/neighclova/service/PlaceService.java
@@ -3,10 +3,10 @@ package com.sogonsogon.neighclova.service;
 import com.sogonsogon.neighclova.domain.Place;
 import com.sogonsogon.neighclova.domain.User;
 import com.sogonsogon.neighclova.dto.object.PlaceListItem;
-import com.sogonsogon.neighclova.dto.request.PlaceRequestDto;
-import com.sogonsogon.neighclova.dto.response.GetAllPlaceResponseDto;
-import com.sogonsogon.neighclova.dto.response.GetPlaceResponseDto;
-import com.sogonsogon.neighclova.dto.response.PlaceResponseDto;
+import com.sogonsogon.neighclova.dto.request.place.PlaceRequestDto;
+import com.sogonsogon.neighclova.dto.response.place.GetAllPlaceResponseDto;
+import com.sogonsogon.neighclova.dto.response.place.GetPlaceResponseDto;
+import com.sogonsogon.neighclova.dto.response.place.PlaceResponseDto;
 import com.sogonsogon.neighclova.dto.response.ResponseDto;
 import com.sogonsogon.neighclova.repository.PlaceRepository;
 import com.sogonsogon.neighclova.repository.UserRepository;
